requirements.txt
# Core dependencies
pyzmq==25.1.1
prometheus-client==0.19.0
watchdog==3.0.0
psutil==5.9.6

# Monitoring and logging
python-json-logger==2.0.7
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
pytest-mock==3.12.0

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.2
isort==5.12.0
bandit==1.7.5  # Security linting

# Security
cryptography==41.0.7
certifi==2023.11.17

# Development tools
pre-commit==3.5.0
ipython==8.18.1
ipdb==0.13.13

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Performance profiling
memory-profiler==0.61.0
line-profiler==4.1.1